refactor(web): break terminal hook import cycle
